requirements.txt
fastapi
uvicorn
psycopg[binary]
jinja2
python-dotenv
